requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.8.2
pandas>=2.2.2
matplotlib>=3.8.0
python-dotenv>=1.0.1
openai>=1.40.6
